refactor(aletheia): extract CLI subcommands into commands/ modules
Split the 2245-line main.rs monolith into focused modules under
crates/aletheia/src/commands/:

- commands/health.rs     — health check handler + HealthArgs
- commands/backup.rs     — backup management + BackupArgs
- commands/maintenance.rs — maintenance tasks + Action enum
- commands/tls.rs        — TLS cert generation + Action enum
- commands/credential.rs — credential status/refresh + Action enum
- commands/eval.rs       — eval scenario runner + EvalArgs
- commands/server.rs     — server startup (was `serve()`) + all helpers
- commands/agent_io.rs   — export/import/skills/migrate commands + Args structs

Each subcommand has a typed Args struct (#[derive(clap::Args)]) and a
single run() entry point. The Command enum variants now hold these structs,
making each dispatch arm a one-liner.

main.rs shrinks from 2245 lines to 172 lines: module declarations, the
top-level Cli + Command types, and a compact dispatch match. Unit tests
moved to src/cli_tests.rs.

- Zero behavior changes — all CLI flags, defaults, and help text unchanged
- cargo clippy --workspace --all-targets -- -D warnings clean
- 36/36 tests pass
