+++ crates/aletheia/src/commands/mod.rs
@@ -1,0 +1,10 @@
+//! CLI subcommand handlers — one module per subcommand.
+
+pub mod agent_io;
+pub mod backup;
+pub mod credential;
+pub mod eval;
+pub mod health;
+pub mod maintenance;
+pub mod server;
+pub mod tls;
